toggle part of sideBar class/module
