+++ src/UI/home.js
@@ -122,29 +122,6 @@ const deleteCategory = (e) => {
 }
 // END CATEGORY LIST //
 
-const toggle = () => {
-    const toggle = document.querySelector('.mobile-toggle');
-    const categoriesNav = document.querySelector('.categories-navigation');
-    toggle.addEventListener('click', () => {
-      const visibility = categoriesNav.getAttribute('data-visible');
-      if (visibility ===  'false')
-      {
-        // Expand nav
-        categoriesNav.setAttribute('data-visible', true);
-        // change Ham-Btn when aria is expanaded
-        toggle.setAttribute('aria-expanded', true);
-      }else
-      {
-        categoriesNav.setAttribute('data-visible', false);
-        toggle.setAttribute('aria-expanded', false);
-      }
-      
-    })
-    
-
-  }
-  
-
 const hero = () => {
   const hero = document.createElement('div');
   hero.classList.add('container', 'container-hero');
@@ -201,7 +178,7 @@ const categoryBtn = (e) => {
 
 export {header,
     hero, 
-    toggle,
+    
     categoryBtn,
     addCategoryPopUp,
     showAddCategoryPanel,
