+++ src/UI/sideBar.js
@@ -1,0 +1,21 @@
+export default class sideBar {
+  static toggle () {
+    const toggle = document.querySelector('.mobile-toggle');
+    const categoriesNav = document.querySelector('.categories-navigation');
+    toggle.addEventListener('click', () => {
+      const visibility = categoriesNav.getAttribute('data-visible');
+      if (visibility ===  'false')
+      {
+        // Expand nav
+        categoriesNav.setAttribute('data-visible', true);
+        // change Ham-Btn when aria is expanaded
+        toggle.setAttribute('aria-expanded', true);
+      }else
+      {
+        categoriesNav.setAttribute('data-visible', false);
+        toggle.setAttribute('aria-expanded', false);
+      }
+      
+    })
+  }
+}
